Rewriting some of LU code to make it easier to plot
The first cell is some of the previous code, but with an attempt to fit
plotting code to make it work. The function is set up so that it appends
to the tim array 4 times per trial. This is difficult to add visual cues
as I couldn't find a way to differentiate between which exponent of n is
being plotted.
The second cell is my attempt at rewriting the code so that it plots
time in seconds vs. n. It should work but takes absurdly long for some
reason.